mcp/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
ruff==0.3.3
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
mcp[cli]==1.9.3
httpx==0.28.1